# IoT_Fermented_Food/auth-ms/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth MS API",
        "version": "1.0.0",
        "description": "API documentation for the Auth microservice"
    },
    "paths": {
        "/login": {
            "post": {
                "summary": "Login user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns true if successful, false otherwise"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns true if successful, false otherwise"
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "summary": "Refresh token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Credential"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns true if successful, false otherwise"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service status"
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "summary": "Prometheus metrics",
                "responses": {
                    "200": {
                        "description": "Metrics data"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "Credential": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "newRefreshToken": {
                        "type": "string"
                    }
                }
            }
        }
    }
}